app/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
isort==5.12.0
packaging==23.1
platformdirs==3.10.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
yapf==0.40.1
zipp==3.17.0
